+++ projects/js-packages/critical-css-gen/src/browser-interface-iframe.ts
@@ -185,7 +185,7 @@ export class BrowserInterfaceIframe extends BrowserInterface {
 					// Check HTTP status code first.
 					const is404 = await this.is404Page( fullUrl );
 					if ( is404 ) {
-						throw new HttpError( { url, code: 404 } );
+						throw new HttpError( { url: fullUrl, code: 404 } );
 					}
 
 					// Verify the inner document is readable.