+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/seo-enhancer/seo-enhancer.tsx
@@ -1,6 +1,7 @@
 /**
  * External dependencies
  */
+import { getAllBlocks } from '@automattic/jetpack-ai-client';
 import { useAnalytics } from '@automattic/jetpack-shared-extension-utils';
 import {
 	BaseControl,
@@ -10,13 +11,15 @@ import {
 	CheckboxControl,
 } from '@wordpress/components';
 import { useSelect, useDispatch } from '@wordpress/data';
-import { useCallback } from '@wordpress/element';
+import { store as editorStore } from '@wordpress/editor';
+import { useCallback, useMemo } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import debugFactory from 'debug';
 /**
  * Internal dependencies
  */
 import { FEATURE_LABELS, FEATURES } from './constants';
+import { SeoEnhancerTaskList } from './seo-enhancer-task-list';
 import { store } from './store';
 import { useSeoModuleSettings } from './use-seo-module-settings';
 import { useSeoRequests } from './use-seo-requests';
@@ -41,7 +44,18 @@ export function SeoEnhancer( {
 
 		return isBusy || isAnyImageBusy;
 	}, [] );
+
 	const enabledFeatures = useSelect( select => select( store ).getEnabledFeatures(), [] );
+	const blocks = useSelect( select => select( editorStore ).getBlocks(), [] );
+
+	const imageBlocks = useMemo( () => {
+		return blocks.length
+			? getAllBlocks().filter(
+					block => block.name === 'core/image' && block.attributes.url && ! block.attributes.alt
+			  )
+			: [];
+	}, [ blocks ] );
+
 	const { setFeatureEnabled } = useDispatch( store );
 
 	const { updateSeoData } = useSeoRequests();
@@ -96,40 +110,33 @@ export function SeoEnhancer( {
 							) }
 						/>
 					) }
-				</BaseControl>
-			</PanelRow>
-			<PanelRow className="jetpack-seo-sidebar__feature-section">
-				<BaseControl __nextHasNoMarginBottom={ true }>
 					{ ( ! isEnabled || disableAutoEnhance ) && (
 						<div className="feature-checkboxes-container">
-							{ FEATURES.map( feature => (
-								<CheckboxControl
-									key={ feature }
-									label={ FEATURE_LABELS[ feature ] }
-									checked={ enabledFeatures.includes( feature ) }
-									onChange={ () => toggleFeature( feature ) }
-									__nextHasNoMarginBottom={ true }
-									disabled={ isLoading }
-									className={ isLoading ? 'is-disabled' : '' }
-								/>
-							) ) }
+							{ FEATURES.map( feature => {
+								const extraLabel =
+									feature === 'images-alt-text' && imageBlocks.length > 0
+										? ` (${ imageBlocks.length })`
+										: '';
+
+								return (
+									<CheckboxControl
+										key={ feature }
+										label={ FEATURE_LABELS[ feature ] + extraLabel }
+										checked={ enabledFeatures.includes( feature ) }
+										onChange={ () => toggleFeature( feature ) }
+										__nextHasNoMarginBottom={ true }
+										disabled={ isLoading }
+										className={ isLoading ? 'is-disabled' : '' }
+									/>
+								);
+							} ) }
 						</div>
 					) }
 					{ isEnabled && ! disableAutoEnhance && (
-						<div className="jetpack-seo-sidebar__feature-list-container">
-							{ enabledFeatures.length > 0 ? (
-								<>
-									<p>{ __( "We'll auto-generate:", 'jetpack' ) }</p>
-									<ul className="jetpack-seo-sidebar__feature-list">
-										{ enabledFeatures.map( feature => (
-											<li key={ feature }>{ FEATURE_LABELS[ feature ] }</li>
-										) ) }
-									</ul>
-								</>
-							) : (
-								<p>{ __( 'No features selected to auto-generate', 'jetpack' ) }</p>
-							) }
-						</div>
+						<SeoEnhancerTaskList
+							isPrePublish={ placement === 'jetpack-prepublish-sidebar' }
+							imageBlocks={ imageBlocks }
+						/>
 					) }
 				</BaseControl>
 			</PanelRow>