+++ projects/js-packages/shared-extension-utils/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.18.10] - 2025-04-02
+### Changed
+- Update package dependencies. [#42809]
+
 ## [0.18.9] - 2025-04-01
 ### Changed
 - Update dependencies. [#42328]
@@ -597,6 +601,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Core: prepare utility for release
 
+[0.18.10]: https://github.com/Automattic/jetpack-shared-extension-utils/compare/0.18.9...0.18.10
 [0.18.9]: https://github.com/Automattic/jetpack-shared-extension-utils/compare/0.18.8...0.18.9
 [0.18.8]: https://github.com/Automattic/jetpack-shared-extension-utils/compare/0.18.7...0.18.8
 [0.18.7]: https://github.com/Automattic/jetpack-shared-extension-utils/compare/0.18.6...0.18.7