+++ projects/packages/analyzer/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.1] - 2025-04-02
+### Changed
+- Code: Use function-style exit() and die() with a default status code of 0. [#41167]
+
+### Fixed
+- Code: Prevent dynamic class properties. [#41857]
+- Fixes a bug in Analyzer tring to call toCodeString() in Variable class. [#40855]
+
 ## [3.0.0] - 2024-12-04
 ### Changed
 - Updated dependencies. [#40286]
@@ -108,6 +116,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Jetpack code analyzer
 
+[3.0.1]: https://github.com/Automattic/jetpack-analyzer/compare/v3.0.0...v3.0.1
 [3.0.0]: https://github.com/Automattic/jetpack-analyzer/compare/v2.0.0...v3.0.0
 [2.0.0]: https://github.com/Automattic/jetpack-analyzer/compare/v1.7.3...v2.0.0
 [1.7.3]: https://github.com/Automattic/jetpack-analyzer/compare/v1.7.2...v1.7.3