+++ projects/packages/google-analytics/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.1] - 2025-04-02
+### Changed
+- Internal updates.
+
 ## [0.3.0] - 2024-12-04
 ### Removed
 - General: Update minimum PHP version to 7.2. [#40147]
@@ -39,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Copy the code from the Jetpack module into the package. [#37184]
 - Migrate unit tests from the Jetpack module. [#37246]
 
+[0.3.1]: https://github.com/Automattic/jetpack-google-analytics/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/Automattic/jetpack-google-analytics/compare/v0.2.5...v0.3.0
 [0.2.5]: https://github.com/Automattic/jetpack-google-analytics/compare/v0.2.4...v0.2.5
 [0.2.4]: https://github.com/Automattic/jetpack-google-analytics/compare/v0.2.3...v0.2.4